dedup ParseImageReferences
